requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
requests==2.32.3
python-dateutil==2.9.0.post0
pydantic==2.10.4
